Move BABA tests to test_dwpc.py
